+++ src/shp_move.cpp
@@ -150,12 +150,14 @@ void Shp_move::finalize()
 
   view().push_undo_delta(std::make_unique<Shape_geom_delta>(std::move(changes)));
   reset();
+  restore_operation_selection_();
 }
 
 void Shp_move::cancel()
 {
   operation_shps_cancel_();
   reset();
+  restore_operation_selection_();
 }
 
 void Shp_move::reset()